feat(wealth-library): phase 0 — extract ContentBody standalone
Extracts the wealth_content document body (Investment Outlooks,
Flash Reports, Manager Spotlights) into
lib/components/library/readers/ContentBody.svelte. Strict prop
contract: only `id`. Optional "netz:content-actor" context
gates the approval UI so LibraryPreviewPane can render content
without showing the Approve button to read-only hosts.

The legacy route at content/[id] becomes a thin shell:
PageHeader (title + breadcrumbs) plus <ContentBody {id} /> under
the route data branches. The +page.server.ts loader is
unchanged.
